+++ web/src/api/namespace.ts
@@ -1,5 +1,7 @@
 import { ProjectAnnotation, Project } from '../../types';
+import YAML from 'yaml';
 import axios from 'axios';
+import { readString } from 'react-papaparse';
 
 const API_HOST = import.meta.env.VITE_API_HOST || '';
 const API_BASE = `${API_HOST}/api/v1`;
@@ -22,6 +24,8 @@ export interface PaginationParams {
   offset?: number;
   limit?: number;
   search?: string;
+  orderBy?: string;
+  order?: 'asc' | 'desc';
 }
 
 export interface ProjectSubmissionResponse {
@@ -46,7 +50,7 @@ export const getNamespaceInfo = (namespace: string, token: string | null = null)
 export const getNamespaceProjects = (
   namespace: string,
   token: string | null = null,
-  { search, offset, limit }: PaginationParams,
+  { search, offset, limit, orderBy, order }: PaginationParams,
 ) => {
   // construct query based on search, offset, and limit
   const query = new URLSearchParams();
@@ -58,6 +62,16 @@ export const getNamespaceProjects = (
   }
   if (limit) {
     query.set('limit', limit.toString());
+  }
+  if (orderBy) {
+    query.set('order_by', orderBy);
+  }
+  if (order) {
+    if (order === 'asc') {
+      query.set('order_desc', 'false');
+    } else {
+      query.set('order_desc', 'true');
+    }
   }
   const url = `${API_BASE}/namespaces/${namespace}/projects?${query.toString()}`;
   if (!token) {
@@ -77,6 +91,7 @@ export const submitProjectFiles = (
     is_private,
     description,
     files,
+    pep_schema,
   }: {
     namespace: string;
     project_name: string;
@@ -84,6 +99,7 @@ export const submitProjectFiles = (
     is_private?: boolean;
     description?: string;
     files: FileList;
+    pep_schema: string;
   },
   token: string,
 ) => {
@@ -95,6 +111,7 @@ export const submitProjectFiles = (
   formData.append('tag', tag || 'default');
   formData.append('is_private', is_private?.toString() || 'false');
   formData.append('description', description || '');
+  formData.append('pep_schema', pep_schema);
 
   // attach files
   for (let i = 0; i < files.length; i++) {
@@ -120,8 +137,9 @@ export const submitProjectJSON = (
     tag,
     is_private,
     description,
-    sample_table,
     config,
+    pep_schema,
+    sample_table,
   }: {
     namespace: string;
     project_name: string;
@@ -130,10 +148,37 @@ export const submitProjectJSON = (
     description?: string;
     sample_table: string;
     config: string;
+    pep_schema: string;
   },
   token: string,
 ) => {
   const url = `${API_BASE}/namespaces/${namespace}/projects/json`;
+
+  // syncronously parse sample table
+  // @ts-ignore
+  const sample_table_json = readString(sample_table, { header: true }).data;
+
+  // json to "sample_dict", whihc has columns as keys to
+  // subobjects and row indexs as subkeys in those subobjects
+  // with the values as the values
+  const sample_dict: {
+    [key: string]: {
+      [key: number]: string;
+    };
+  } = {};
+  for (let i = 0; i < sample_table_json.length; i++) {
+    const row = sample_table_json[i];
+    for (const key in row) {
+      if (key in sample_dict) {
+        sample_dict[key][i] = row[key];
+      } else {
+        sample_dict[key] = {};
+        sample_dict[key][i] = row[key];
+      }
+    }
+  }
+
+  const config_json = YAML.parse(config);
   return axios
     .post<ProjectSubmissionResponse>(
       url,
@@ -141,7 +186,9 @@ export const submitProjectJSON = (
         pep_dict: {
           name: project_name,
           description: description || '',
-          config: config,
+          config: config_json,
+          sample_dict: sample_dict,
+          pep_schema: pep_schema,
         },
         is_private: is_private || false,
         tag: tag || 'default',