+++ web/src/components/forms/edit-project-meta.tsx
@@ -1,29 +1,45 @@
-import { FC } from 'react';
-import { SubmitHandler, useForm } from 'react-hook-form';
+import { FC, useEffect } from 'react';
+import { SubmitHandler, useForm, Controller } from 'react-hook-form';
 import { editProjectMetadata } from '../../api/project';
 import { useSession } from '../../hooks/useSession';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'react-hot-toast';
 import { useProject } from '../../hooks/queries/useProject';
+import { SchemaDropdown } from './components/schemas-databio-dropdown';
+import { MarkdownEditor } from '../markdown/edit';
+import { AxiosError } from 'axios';
 
 interface Props {
   namespace: string;
   name: string;
   tag: string;
+  onSuccessfulSubmit?: () => void;
+  onFailedSubmit?: () => void;
+  onCancel?: () => void;
 }
 
 interface FormValues extends Props {
   description: string;
   isPrivate: boolean;
+  pep_schema: string;
 }
 
-export const ProjectMetaEditForm: FC<Props> = ({ namespace, name, tag }) => {
+export const ProjectMetaEditForm: FC<Props> = ({
+  namespace,
+  name,
+  tag,
+  onSuccessfulSubmit = () => {},
+  onFailedSubmit = () => {},
+  onCancel = () => {},
+}) => {
   const { jwt } = useSession();
   const { data: projectInfo } = useProject(namespace, name, tag, jwt);
   const {
     register,
     handleSubmit,
     watch,
+    control,
+    setValue,
     reset: resetForm,
     formState: { isValid, isDirty },
   } = useForm<FormValues>({
@@ -32,6 +48,7 @@ export const ProjectMetaEditForm: FC<Props> = ({ namespace, name, tag }) => {
       description: projectInfo?.description || '',
       isPrivate: projectInfo?.is_private,
       tag: tag,
+      pep_schema: projectInfo?.pep_schema || 'pep/2.1.0',
     },
   });
   const newTag = watch('tag');
@@ -54,14 +71,21 @@ export const ProjectMetaEditForm: FC<Props> = ({ namespace, name, tag }) => {
       );
       toast.success('Project metadata updated successfully.');
       queryClient.invalidateQueries([namespace, name, tag]);
+      onSuccessfulSubmit();
 
       // if newTag or newName is different, redirect to new project
       if (newTag !== tag || newName !== name) {
         window.location.href = `/${namespace}/${newName}?tag=${newTag}`;
       }
     },
-    onError: (error) => {
+    onError: (error: AxiosError) => {
+      // check for axios 401
+      if (error.response?.status === 401) {
+        toast.error('You are not authorized to edit this project.');
+        return;
+      }
       toast.error(`There was an error updated project metadata: ${error}`);
+      onFailedSubmit();
     },
   });
 
@@ -96,6 +120,27 @@ export const ProjectMetaEditForm: FC<Props> = ({ namespace, name, tag }) => {
         </div>
       </div>
       <div className="mb-3">
+        <label htmlFor="schema-tag" className="form-label">
+          Schema
+        </label>
+        <div>
+          <Controller
+            control={control}
+            name="pep_schema"
+            render={({ field: { onChange, value } }) => (
+              <SchemaDropdown
+                value={value}
+                onChange={(schema) => {
+                  setValue('pep_schema', schema, {
+                    shouldDirty: true,
+                  });
+                }}
+              />
+            )}
+          />
+        </div>
+      </div>
+      <div className="mb-3">
         <label htmlFor="project-tag" className="form-label">
           Project Tag
         </label>
@@ -114,15 +159,28 @@ export const ProjectMetaEditForm: FC<Props> = ({ namespace, name, tag }) => {
         <label htmlFor="project-description" className="form-label">
           Project Description
         </label>
-        <textarea
-          {...register('description')}
-          placeholder="Project description"
-          className="form-control"
-          id="project-description"
-          rows={3}
-        ></textarea>
+        <Controller
+          control={control}
+          name="description"
+          render={({ field }) => (
+            <MarkdownEditor
+              name="description"
+              value={field.value}
+              onChange={(value) => {
+                field.onChange(value);
+              }}
+            />
+          )}
+        />
       </div>
-      <button onClick={() => resetForm()} type="button" className="btn btn-outline-dark me-1">
+      <button
+        onClick={() => {
+          onCancel();
+          resetForm();
+        }}
+        type="button"
+        className="btn btn-outline-dark me-1"
+      >
         Cancel
       </button>
       <button